+++ src/world_gens/world_generators.hpp
@@ -5,11 +5,11 @@
 #include <string>
 
 namespace worldgens {
-    extern const double G;
+    extern const float G;
 
     class WorldGenerator {
         public:
-            virtual void Generate(float*& particleArray, int n) = 0;
+            virtual void Generate(float*& particleArray, unsigned int n) = 0;
             virtual std::string ToString() const = 0;
             virtual std::unique_ptr<WorldGenerator> Clone() const = 0;
             bool IsSameType(const WorldGenerator& other) const;
@@ -20,35 +20,35 @@ namespace worldgens {
 
     class GalaxyGenerator : public WorldGenerator {
         public:
-            void Generate(float*& particleArray, int n) override;
+            void Generate(float*& particleArray, unsigned int n) override;
             std::string ToString() const override;
             std::unique_ptr<WorldGenerator> Clone() const override;
     };
 
     class TwoGalaxiesGenerator : public WorldGenerator {
         public:
-            void Generate(float*& particleArray, int n) override;
+            void Generate(float*& particleArray, unsigned int n) override;
             std::string ToString() const override;
             std::unique_ptr<WorldGenerator> Clone() const override;
     };
 
     class SphereGenerator : public WorldGenerator {
         public:
-            void Generate(float*& particleArray, int n) override;
+            void Generate(float*& particleArray, unsigned int n) override;
             std::string ToString() const override;
             std::unique_ptr<WorldGenerator> Clone() const override;
     };
 
     class SphereShellGenerator : public WorldGenerator {
         public:
-            void Generate(float*& particleArray, int n) override;
+            void Generate(float*& particleArray, unsigned int n) override;
             std::string ToString() const override;
             std::unique_ptr<WorldGenerator> Clone() const override;
     };
 
     class TwoSpheresGenerator : public WorldGenerator {
         public:
-            void Generate(float*& particleArray, int n) override;
+            void Generate(float*& particleArray, unsigned int n) override;
             std::string ToString() const override;
             std::unique_ptr<WorldGenerator> Clone() const override;
     };
